implement SearchFragment
